refactor: absorb text styling helpers into recipe module
